Refactor mood controller page.
refactor components in mood controller page.
