+++ src/components/MoodBrightnessForm.jsx
@@ -1,56 +1,29 @@
 import React from 'react';
 
-import styled from '@emotion/styled';
-
-import Form from '../styles/Form';
-import RadioWrap from '../styles/RadioWrap';
-
 import {
-  MoodBackground,
-  MoodBackgroundHover,
-  MoodBackgroundActive,
-  MoodBackgroundSelected,
-} from '../assets/images';
-
-const Label = styled.label(({ checked }) => ({
-  fontSize: '32px',
-  fontWeight: 'bold',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  margin: '0 17px',
-  width: '32vh',
-  height: '13vh',
-  background: `url(${checked ? MoodBackgroundSelected : MoodBackground}) no-repeat`,
-  backgroundSize: 'contain',
-
-  '&:hover': {
-    background: `url(${checked ? MoodBackgroundSelected : MoodBackgroundHover}) no-repeat`,
-    backgroundSize: 'contain',
-  },
-
-  '&:active': {
-    background: `url(${MoodBackgroundActive}) no-repeat`,
-    backgroundSize: 'contain',
-
-    '& p': {
-      transform: 'translateY(10px)',
-    },
-  },
-}));
+  Form,
+  RadioWrap,
+  Label,
+} from '../styles/moodSeletForm';
 
 const MoodBrightnessForm = React.memo(({ onChange, field }) => {
-  const formValues = [{ id: 'happy', text: '밝은' }, { id: 'dark', text: '어두운' }];
+  const formValues = [
+    { id: 'happy', text: '밝은', value: 'happy' },
+    { id: 'dark', text: '어두운', value: 'dark' },
+    { id: 'no-brightness', text: '모르겠어요', value: 'none' },
+  ];
 
   return (
     <Form>
       <h2>밝은 느낌이 좋으세요? 아니면 어두운 느낌이 좋으세요?</h2>
       <RadioWrap>
-        {formValues.map(({ id, text }) => (
+        {formValues.map(({
+          id, text, value,
+        }) => ((
           <Label
             htmlFor={id}
             key={id}
-            checked={field === id}
+            checked={field === value}
           >
             <p>
               {text}
@@ -59,28 +32,13 @@ const MoodBrightnessForm = React.memo(({ onChange, field }) => {
               type="radio"
               id={id}
               name="brightness"
-              value={id}
+              value={value}
               onChange={onChange}
             />
           </Label>
-        ))}
-        <Label
-          htmlFor="no-brightness"
-          checked={field === 'none'}
-        >
-          모르겠어요
-          <input
-            type="radio"
-            id="no-brightness"
-            name="brightness"
-            value="none"
-            onChange={onChange}
-            placeholder="모르겠어요"
-          />
-        </Label>
+        )))}
       </RadioWrap>
     </Form>
-
   );
 });
