+++ src/components/MoodEnergyForm.jsx
@@ -1,56 +1,29 @@
 import React from 'react';
 
-import styled from '@emotion/styled';
-
-import Form from '../styles/Form';
-import RadioWrap from '../styles/RadioWrap';
-
 import {
-  MoodBackground,
-  MoodBackgroundHover,
-  MoodBackgroundActive,
-  MoodBackgroundSelected,
-} from '../assets/images';
-
-const Label = styled.label(({ checked }) => ({
-  fontSize: '32px',
-  fontWeight: 'bold',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  margin: '0 17px',
-  width: '32vh',
-  height: '13vh',
-  background: `url(${checked ? MoodBackgroundSelected : MoodBackground}) no-repeat`,
-  backgroundSize: 'contain',
-
-  '&:hover': {
-    background: `url(${checked ? MoodBackgroundSelected : MoodBackgroundHover}) no-repeat`,
-    backgroundSize: 'contain',
-  },
-
-  '&:active': {
-    background: `url(${MoodBackgroundActive}) no-repeat`,
-    backgroundSize: 'contain',
-
-    '& p': {
-      transform: 'translateY(10px)',
-    },
-  },
-}));
+  Form,
+  RadioWrap,
+  Label,
+} from '../styles/moodSeletForm';
 
 const MoodEnergyForm = React.memo(({ onChange, field }) => {
-  const formValues = [{ id: 'calm', text: '차분한' }, { id: 'uplifting', text: '신나는' }];
+  const formValues = [
+    { id: 'calm', text: '차분한', value: 'calm' },
+    { id: 'uplifting', text: '신나는', value: 'uplifting' },
+    { id: 'no-energy', text: '모르겠어요', value: 'none' },
+  ];
 
   return (
     <Form>
       <h2>차분하고 싶으세요 아니면 신나고 싶으세요?</h2>
       <RadioWrap>
-        {formValues.map(({ id, text }) => (
+        {formValues.map(({
+          id, text, value,
+        }) => ((
           <Label
             htmlFor={id}
             key={id}
-            checked={field === id}
+            checked={field === value}
           >
             <p>
               {text}
@@ -59,24 +32,11 @@ const MoodEnergyForm = React.memo(({ onChange, field }) => {
               type="radio"
               id={id}
               name="energy"
-              value={id}
+              value={value}
               onChange={onChange}
             />
           </Label>
-        ))}
-        <Label
-          htmlFor="no-energy"
-          checked={field === 'none'}
-        >
-          모르겠어요
-          <input
-            type="radio"
-            id="no-energy"
-            name="energy"
-            value="none"
-            onChange={onChange}
-          />
-        </Label>
+        )))}
       </RadioWrap>
     </Form>
   );
