--- src/styles/Form.js
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-const Form = styled.form({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  marginTop: '20px',
-});
-
-export default Form;
