--- src/styles/RadioWrap.js
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-const RadioWrap = styled.div({
-  display: 'flex',
-  flex: 'row',
-  margin: '10px',
-  padding: '10px 20px',
-  border: '1px solid #000',
-  borderRadius: '5px',
-});
-
-export default RadioWrap;
